Refactor SessionManager and add HttpSessionManager
This commit refactors the existing Redis-based SessionManager and introduces a new HttpSessionManager, allowing client login type strategy to be configurable which use HttpSessionManager.

Key changes:
1.  Extracted `SessionManager` interface from the original `SessionManager` class.
    - Defines methods for session creation, retrieval, destruction, renewal, and add ticket manger interface for ticket management.

2.  Renamed the original `SessionManager` to `RedisSessionManager` and made it implement the new `SessionManager` interface.
    - This class continues to use Redis for session storage.

3.  Introduced `HttpSessionManager` which also implements `SessionManager`.
    - This implementation validates sessions by making HTTP calls to `/openapi/userInfo` (for getSession) and `/openapi/logout` (for destroySession) endpoints.
    - `create()` and ticket-related methods throw `NotImplementedException` as these operations are not handled via HTTP delegation.
    - `renew()` is a no-op.
    - Requires `bella.login.openapi-base` and cookie name (via `SessionProperty`) to be configured.

4.  Expanded `bella.login.type` to include a `client` type to control which `SessionManager` implementation is used.

5.  Modified `BellaLoginConfiguration` to:
    - Conditionally instantiate either `RedisSessionManager` or `HttpSessionManager` based on the `bella.login.type` property.
    - Provide necessary dependencies (RestTemplate, base URL, SessionProperty) to `HttpSessionManager`.
    - Ensure `RedisSessionManager` is configured with its dependencies as before.

6.  Added comprehensive unit tests for `HttpSessionManager` to verify its logic, including HTTP call, cookie extraction, and handling of various response scenarios.

7.  Verified that existing consumers within the `api/server` module (like `LoginFilter`) use the `SessionManager` interface, allowing them to work seamlessly with either implementation.

This change enables services using the `bella-openapi-spi` to choose between direct Redis session management or delegating session validation to a central `bella-openapi` server via HTTP calls, enhancing flexibility and decoupling session storage requirements for client services.
