+++ api/sdk/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <groupId>io.github.bella-top</groupId>
         <artifactId>bella-openapi</artifactId>
-        <version>1.0.3</version>
+        <version>1.1.0</version>
     </parent>
 
     <artifactId>sdk</artifactId>
