+++ api/server/pom.xml
@@ -4,7 +4,7 @@
     <parent>
         <groupId>io.github.bella-top</groupId>
         <artifactId>bella-openapi</artifactId>
-        <version>1.0.3</version>
+        <version>1.1.0</version>
     </parent>
     <artifactId>server</artifactId>
     <name>server</name>
