+++ api/spi/src/main/java/com/ke/bella/openapi/login/session/TicketManager.java
@@ -1,0 +1,9 @@
+package com.ke.bella.openapi.login.session;
+
+public interface TicketManager {
+    void saveTicket(String ticket);
+
+    boolean isValidTicket(String ticket);
+
+    void removeTicket(String ticket);
+}
